feat: basic_routes
