+++ src/App.jsx
@@ -1,38 +1,17 @@
+import { Route, Routes } from 'react-router-dom';
 import './App.scss';
-import CategoryList from './components/category-list/categories-component';
+import Home from './routes/Home/home-component'
+import Navigation from './routes/NavBar/navigation-component';
+import Shop from './routes/Shop/shop-component';
 
 const App = () => {
-
-  const categories = [
-    {
-      "id": 1,
-      "title": "Hats",
-      "imageUrl": "https://i.ibb.co/cvpntL1/hats.png"
-    },
-    {
-      "id": 2,
-      "title": "Jackets",
-      "imageUrl": "https://i.ibb.co/px2tCc3/jackets.png"
-    },
-    {
-      "id": 3,
-      "title": "Sneakers",
-      "imageUrl": "https://i.ibb.co/0jqHpnp/sneakers.png"
-    },
-    {
-      "id": 4,
-      "title": "Women's",
-      "imageUrl": "https://i.ibb.co/GCCdy8t/womens.png"
-    },
-    {
-      "id": 5,
-      "title": "Men's",
-      "imageUrl": "https://i.ibb.co/R70vBrQ/men.png"
-    }
-  ]
-
   return (
-    <CategoryList categories={categories}/>
+    <Routes>
+      <Route path='/' element={<Navigation />}>
+        <Route index element={<Home/>} />
+        <Route path='shop' element={<Shop />} />
+      </Route>
+    </Routes>
   )
 }
