+++ src/index.scss
@@ -1,6 +1,6 @@
 *{
-  margin: 0;
-  padding: 0;
+  margin: 0px;
+  padding: 0px;
   border: 0px;
   outline: none;
   appearance: none;
@@ -10,6 +10,7 @@
 body {
   margin: 0;
 font-family: 'Lora', sans-serif;
+text-decoration: none;
 }
 
 code {
