+++ src/routes/Shop/shop-component.jsx
@@ -1,0 +1,7 @@
+const Shop = () => {
+    return(
+      <div>Shop</div>
+    )
+  }
+
+  export default Shop;
